Standardize on test harness and layout for pkg/apis
