+++ pkg/apis/azure/v1alpha1/types_test.go
@@ -17,13 +17,47 @@ limitations under the License.
 package v1alpha1
 
 import (
+	"context"
+	"log"
 	"testing"
 
+	"github.com/crossplaneio/crossplane/pkg/test"
 	"github.com/onsi/gomega"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
+
+const (
+	namespace     = "default"
+	name          = "test-provider"
+	secretDataKey = "credentials.json"
+)
+
+var (
+	cfg *rest.Config
+	c   client.Client
+	ctx = context.TODO()
+)
+
+func TestMain(m *testing.M) {
+	err := SchemeBuilder.AddToScheme(scheme.Scheme)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	t := test.NewTestEnv(namespace, test.CRDs())
+	cfg = t.Start()
+
+	if c, err = client.New(cfg, client.Options{Scheme: scheme.Scheme}); err != nil {
+		log.Fatal(err)
+	}
+
+	t.StopAndExit(m.Run())
+}
 
 func TestStorageProvider(t *testing.T) {
 	key := types.NamespacedName{Name: name, Namespace: namespace}
