--- pkg/apis/core/v1alpha1/v1alpha1_suite_test.go
@@ -1,55 +1,0 @@
-/*
-Copyright 2018 The Crossplane Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package v1alpha1
-
-import (
-	"log"
-	"testing"
-
-	"github.com/crossplaneio/crossplane/pkg/test"
-	"golang.org/x/net/context"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-const (
-	namespace = "default"
-	name      = "test-instance"
-)
-
-var (
-	ctx = context.TODO()
-	cfg *rest.Config
-	c   client.Client
-)
-
-func TestMain(m *testing.M) {
-	err := SchemeBuilder.AddToScheme(scheme.Scheme)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	t := test.NewTestEnv(namespace, test.CRDs())
-	cfg = t.Start()
-
-	if c, err = client.New(cfg, client.Options{Scheme: scheme.Scheme}); err != nil {
-		log.Fatal(err)
-	}
-
-	t.StopAndExit(m.Run())
-}
